+++ packages/tswebextension/CHANGELOG.md
@@ -11,8 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.7] - 2023-07-21
 
 ### Changed
-- Updated `@adguard/tsurlfilter` to `v2.1.4`.
+- Updated `@adguard/tsurlfilter` to `v2.1.5`.
 - Updated `@adguard/scriptlets` to `v1.9.57`.
+
+## Fixed
+- Duplicate `eventId` of filtering events.
 
 
 ## [0.3.6] - 2023-07-11